Time: 8 ms (8.9%), Space: 95.1 MB (62.65%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.cpp
@@ -19,7 +19,7 @@ public:
             if(mid2 - 1 >= 0) l2 = nums2[mid2 - 1];
             if(l1 <= r2 && l2 <= r1){
                 if(n % 2 == 1) return max(l1, l2);
-                return (double)(((double)(max(l1, l2)+min(r1, r2)))/2.0);
+                return ((double)(max(l1, l2)+min(r1, r2)))/2.0;
             }
             else if(l1 > r2) high = mid1 - 1;
             else low = mid1 + 1;
